pyproject.toml
[project]
authors = [{ name = "SoulMelody", email = "yjxrtzyx@gmail.com" }]
version = "0.1.1"
requires-python = ">=3.10,<3.14"
license = { text = "MIT" }
name = "anyf0"
description = "Use VocalShifter to visualize f0 curve instead of matplotlib"
keywords = []
classifiers = []
dependencies = [
    "construct>=2.10.70",
    "libf0>=1.1.1",
    "numba>=0.63.1",
    "praat-parselmouth>=0.4.7",
    "pretty-midi>=0.2.11",
    "pyworld-prebuilt>=0.3.5.post2",
    "tgt>=1.5",
]

[project.optional-dependencies]
cpu = [
    "onnxruntime>=1.23.2",
    "swift-f0>=0.1.2",
]
gpu = [
    "onnxruntime-gpu>=1.23.2",
    "swift-f0>=0.1.2",
]
webgpu = [
    "onnxruntime-webgpu>=1.24.0.dev20251218001",
    "swift-f0>=0.1.2",
]

[dependency-groups]
dev = [
    "jaxtyping>=0.3.7",
    "matplotlib>=3.10.8",
]

[tool.uv]
conflicts = [
    [
        { extra = "cpu" },
        { extra = "gpu" },
        { extra = "webgpu" },
    ],
]

[[tool.uv.index]]
name = "default"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"

[project.scripts]
anyf0 = "anyf0.__main__:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
